Fix small problems due to Django version changing
Previously, email field's max length defaults to 75; and now to 254. Such
changes unnecessarily force us to make new migration even for a change
as trival as white space in models.py. To avoid that, we now pin down
the max length.

The last_login field is out of our control; we will just float with the
default.